add saver products

+++ WarehouseManagementSystem/Task2_UserManagement/WarehouseService.cs
@@ -1,9 +1,11 @@
 using System.Data;
+using Newtonsoft.Json;
 using WarehouseManagementSystem.Shared.Enums;
 using WarehouseManagementSystem.Task1_DataModel.Products;
 using WarehouseManagementSystem.Task3_WarehouseLogic;
 using WarehouseManagementSystem.Task4_BusinessAndSecurity.Validation;
 using WarehouseManagementSystem.Task4_BusinessAndSecurity.Logging;
+using WarehouseManagementSystem.SaverLoader;
 
 namespace WarehouseManagementSystem.Task2_UserManagement;
 
@@ -28,13 +30,47 @@ public class WarehouseService
         _userService = userService;
     }
 
-    public void AddUser(string name, string password, UserRole role)
+    public void AuthLog(User user)
     {
+        _logger.Info($"User: {user.UserName} logged in. {DateTime.Now}");
+    }
+    public void AddUser(User user ,string name, string password, UserRole role)
+    {
+        var userValidationResult = _usersValidator.Validate(user);
+        if (!userValidationResult.IsValid)
+        {
+            Console.WriteLine("This user is blocked");
+            _logger.Warn($"Blocked User:{user.UserName} tried remove product. {DateTime.Now}");
+            Console.ReadKey();
+            return;
+        }
+        if (!AccessControl.CanManageUsers(user))
+        {
+            Console.WriteLine("This user is not allowed manage users");
+            _logger.Warn($"User: {user.UserName} is not to add users. {DateTime.Now}");
+            Console.ReadKey();
+            return;
+        }
         _userService.Register(name, password, role);
     }
 
-    public void RemoveUser(Guid userId)
+    public void RemoveUser(User user ,Guid userId)
     {
+        var userValidationResult = _usersValidator.Validate(user);
+        if (!userValidationResult.IsValid)
+        {
+            Console.WriteLine("This user is blocked");
+            _logger.Warn($"Blocked User:{user.UserName} tried remove user. {DateTime.Now}");
+            Console.ReadKey();
+            return;
+        }
+        if (!AccessControl.CanManageUsers(user))
+        {
+            Console.WriteLine("This user is not allowed manage users");
+            _logger.Warn($"User: {user.UserName} is not allowed to remove users. {DateTime.Now}");
+            Console.ReadKey();
+            return;
+        }
         _userService.DeleteUser(userId);
     }
 
@@ -55,7 +91,7 @@ public class WarehouseService
         {
             Console.WriteLine("This user is blocked");
             _logger.Warn($"Blocked User:{user.UserName} tried add product. {DateTime.Now}");
-            
+            Console.ReadKey();
             return;
         }
         var productValidationResult = _productsValidator.Validate(product);
@@ -63,12 +99,14 @@ public class WarehouseService
         {
             Console.WriteLine("Invalid product");
             _logger.Error($"Trying to add invalid product. {DateTime.Now}");
+            Console.ReadKey();
             return;
         }
         if (!AccessControl.CanAddProduct(user))
         {
             Console.WriteLine("This user is not allowed to add product");
             _logger.Warn($"User: {user.UserName} is not allowed to add product. {DateTime.Now}");
+            Console.ReadKey();
             return;
         }
         _warehouse.AddProduct(product);
@@ -78,16 +116,18 @@ public class WarehouseService
     public void RemoveProduct(User user, Guid productId)
     {
         var userValidationResult = _usersValidator.Validate(user);
-        if (!userValidationResult.IsValid)
+        if (userValidationResult.IsValid)
         {
             Console.WriteLine("This user is blocked");
             _logger.Warn($"Blocked User:{user.UserName} tried remove product. {DateTime.Now}");
+            Console.ReadKey();
             return;
         }
         if (!AccessControl.CanAddProduct(user))
         {
             Console.WriteLine("This user is not allowed to remove product");
             _logger.Warn($"User: {user.UserName} is not allowed to remove product. {DateTime.Now}");
+            Console.ReadKey();
             return;
         }
         _warehouse.RemoveProduct(productId);
@@ -97,10 +137,11 @@ public class WarehouseService
     public void BlockUser(User user, Guid userId)
     {
         var userValidationResult = _usersValidator.Validate(user);
-        if (!userValidationResult.IsValid)
+        if (userValidationResult.IsValid)
         {
             Console.WriteLine("This user is blocked");
             _logger.Warn($"Blocked User:{user.UserName} tried to block user. {DateTime.Now}");
+            Console.ReadKey();
             return;
         }
         if(user.Id == userId) {return;}
@@ -109,6 +150,7 @@ public class WarehouseService
         {
             Console.WriteLine("This user is not allowed to manage users");
             _logger.Warn($"User: {user.UserName} is not allowed to block users. {DateTime.Now}");
+            Console.ReadKey();
             return;
         }
         _userService.BlockUser(userId);
@@ -122,6 +164,7 @@ public class WarehouseService
         {
             Console.WriteLine("This user is blocked");
             _logger.Warn($"Blocked User:{user.UserName} tried to unblock user. {DateTime.Now}");
+            Console.ReadKey();
             return;
         }
         if(user.Id == userId) {return;}
@@ -130,6 +173,7 @@ public class WarehouseService
         {
             Console.WriteLine("This user is not allowed to unblock user");
             _logger.Warn($"User: {user.UserName} is not allowed to unblock users. {DateTime.Now}");
+            Console.ReadKey();
             return;
         }
         _userService.BlockUser(userId);
@@ -138,10 +182,11 @@ public class WarehouseService
 
     public void UpdateProduct(User user, Guid id, int choice, double? price, int? quantity)
     {
-        if (!_usersValidator.Validate(user).IsValid)
+        if (_usersValidator.Validate(user).IsValid)
         {
             Console.WriteLine("This user is blocked");
             _logger.Warn($"Blocked User:{user.UserName} tried to update product. {DateTime.Now}");
+            Console.ReadKey();
             return;
         }
 
@@ -149,6 +194,7 @@ public class WarehouseService
         {
             Console.WriteLine("This user is not allowed to add product");
             _logger.Warn($"User: {user.UserName} is not allowed to update product. {DateTime.Now}");
+            Console.ReadKey();
             return;
         }
         switch (choice)
@@ -159,6 +205,7 @@ public class WarehouseService
                 {
                     Console.WriteLine("Price is required");
                     _logger.Error($"Failed to update product price. {DateTime.Now}");
+                    Console.ReadKey();
                     return;
                 }
                 _warehouse.ChangeProductPrice(id , price.Value);
@@ -171,6 +218,7 @@ public class WarehouseService
                 {
                     Console.WriteLine("Quantity is required");
                     _logger.Error($"Failed to update product quantity. {DateTime.Now}");
+                    Console.ReadKey();
                     return;
                 }
                 _warehouse.ChangeProductQuantity(id, quantity.Value);
@@ -182,5 +230,10 @@ public class WarehouseService
                 _logger.Error($"Failed to update product. {DateTime.Now}");
                 break;
         }
+    }
+    public void Save()
+    {
+        _userService.UserSave();
+        Saver.SaveProducts(_warehouse.Products);
     }
 }
